be/requirements.txt
strands-agents>=0.3.0
strands-agents-tools>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
mangum>=0.17.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
sse-starlette>=1.8.0
httpx>=0.27.0
brotli>=1.1.0
curl_cffi>=0.14.0
